yolo_app/server/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
alembic
